move UuidFileHandle to a separate module
